feat/update function to show the diff url

+++ src/main/kotlin/kt/ExampleResource.kt
@@ -5,6 +5,7 @@ import jakarta.ws.rs.POST
 import jakarta.ws.rs.Path
 import jakarta.ws.rs.core.MediaType
 import org.slf4j.LoggerFactory
+import java.math.BigDecimal
 
 @Path("/github-webhook")
 class ExampleResource {
@@ -18,7 +19,7 @@ class ExampleResource {
         val action = payload["action"] as String?
         logger.info("Action: ${action ?: "unknown"}")
         if (action == "opened") {
-            val prNumber = payload["number"] as Int?
+            val prNumber = payload["number"] as BigDecimal?
             val pullRequest = payload["pull_request"] as Map<*, *>?
             val diffUrl = pullRequest?.get("diff_url") as String?
             logger.info("Processing PR # $prNumber - Diff URL: $diffUrl")
